For #14733：Sinking the readValue method of AbstractInventoryDumper into PostgreSQLInventoryDumper class
