+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-dialect/shardingsphere-data-pipeline-postgresql/src/main/java/org/apache/shardingsphere/data/pipeline/postgresql/ingest/PostgreSQLInventoryDumper.java
@@ -31,6 +31,8 @@ import java.sql.SQLException;
  */
 public final class PostgreSQLInventoryDumper extends AbstractInventoryDumper {
     
+    private static final String PG_MONEY_TYPE = "money";
+    
     public PostgreSQLInventoryDumper(final InventoryDumperConfiguration inventoryDumperConfig, final PipelineDataSourceManager dataSourceManager) {
         super(inventoryDumperConfig, dataSourceManager);
     }
@@ -40,5 +42,17 @@ public final class PostgreSQLInventoryDumper extends AbstractInventoryDumper {
         PreparedStatement result = connection.prepareStatement(sql, ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
         result.setFetchSize(1);
         return result;
+    }
+    
+    @Override
+    protected Object readValue(final ResultSet resultSet, final int index) throws SQLException {
+        if (isPgMoneyType(resultSet, index)) {
+            return resultSet.getBigDecimal(index);
+        }
+        return resultSet.getObject(index);
+    }
+    
+    private boolean isPgMoneyType(final ResultSet resultSet, final int index) throws SQLException {
+        return PG_MONEY_TYPE.equalsIgnoreCase(resultSet.getMetaData().getColumnTypeName(index));
     }
 }
